pyproject.toml
[tool.poetry]
name = "abank-marketing-crew"
version = "1.0.0"
description = "Autonomous Multi-Agent Marketing System for ABank using CrewAI"
authors = ["ABank AI Team"]
readme = "README.md"
packages = [{include = "abank_marketing_crew", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.86.0"
crewai-tools = "^0.12.1"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
pyyaml = "^6.0.0"
rich = "^13.9.0"
click = "^8.1.0"
requests = "^2.32.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
flake8 = "^7.1.0"
mypy = "^1.13.0"
isort = "^5.13.0"

[tool.poetry.scripts]
abank-crew = "abank_marketing_crew.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src/abank_marketing_crew --cov-report=html --cov-report=term"
